requirements.txt
torch==2.2.2
numpy<2.0
Flask
Flask-Reuploaded
flask_cors
werkzeug
flask_sqlalchemy
celery
redis
gevent
pandas
openpyxl
scikit-learn
scipy
Pillow
transformers==4.39.3
sentencepiece
spacy
spacy-transformers
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.2/en_core_web_trf-3.7.2-py3-none-any.whl
pymupdf
symspellpy
